1) Additional trace tags

+++ Source/DotNetWorkQueue.Transport.PostgreSQL/ConfigurationExtensions.cs
@@ -72,7 +72,7 @@ namespace DotNetWorkQueue.Transport.PostgreSQL
         /// </remarks>
         public static ushort? GetPriority(this IAdditionalMessageData data)
         {
-            return data.TryGetSetting("PostgreSQLMessageQueuePriority", out dynamic value) ? value : 128;
+            return data.TryGetSetting("PostgreSQLMessageQueuePriority", out dynamic value) ? value : (ushort)128;
         }
     }
